# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Përgjithmonë Photobook Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List current user's projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new photobook project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project with all its pages
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project with pages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"

    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project metadata
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{projectId}/pages:
    post:
      operationId: addProjectPage
      tags: [projects]
      summary: Add a page to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPageInput"
      responses:
        "201":
          description: Page added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPage"
  /projects/{projectId}/pages/{pageId}:
    patch:
      operationId: updateProjectPage
      tags: [projects]
      summary: Update a page's content
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPageUpdate"
      responses:
        "200":
          description: Page updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPage"

    delete:
      operationId: deleteProjectPage
      tags: [projects]
      summary: Delete a page
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{projectId}/auto-save:
    post:
      operationId: autoSaveProject
      tags: [projects]
      summary: Auto-save project state
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoSaveInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{projectId}/generate-pdf:
    post:
      operationId: generateProjectPdf
      tags: [projects]
      summary: Generate print-ready PDF for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfGenerationResponse"
  /projects/{projectId}/pdf-status:
    get:
      operationId: getProjectPdfStatus
      tags: [projects]
      summary: Check PDF generation status
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfGenerationResponse"

  # ─── ORDERS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    Project:
      type: object
      required: [id, userId, title, status, bookSizeId, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        templateId:
          type: ["integer", "null"]
        bookSizeId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [draft, pdf_generating, pdf_ready, ordered]
        pdfUrl:
          type: ["string", "null"]
        shareToken:
          type: ["string", "null"]
        pageCount:
          type: integer
        totalPriceLek:
          type: integer
        frontCoverJson:
          type: ["string", "null"]
          description: JSON-serialized element array for the front cover page, used to render a thumbnail without a separate pages fetch.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [bookSizeId, title]
      properties:
        templateId:
          type: integer
        bookSizeId:
          type: integer
        title:
          type: string
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        bookSizeId:
          type: integer
    ProjectPage:
      type: object
      required: [id, projectId, pageNumber, pageType]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        pageNumber:
          type: integer
        pageType:
          type: string
          enum: [front_cover, inside_cover, inner, back_cover]
        layoutId:
          type: ["integer", "null"]
        contentJson:
          type: string
          description: JSON string of page canvas content
        updatedAt:
          type: string
          format: date-time
    ProjectPageInput:
      type: object
      required: [pageNumber, pageType]
      properties:
        pageNumber:
          type: integer
        pageType:
          type: string
          enum: [front_cover, inside_cover, inner, back_cover]
        layoutId:
          type: integer
        contentJson:
          type: string
    ProjectPageUpdate:
      type: object
      properties:
        layoutId:
          type: integer
        contentJson:
          type: string
        pageNumber:
          type: integer
    ProjectDetail:
      type: object
      required: [id, userId, title, status, bookSizeId, createdAt, updatedAt, pages]
      properties:
        id:
          type: integer
        userId:
          type: integer
        templateId:
          type: ["integer", "null"]
        bookSizeId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [draft, pdf_generating, pdf_ready, ordered]
        pdfUrl:
          type: ["string", "null"]
        shareToken:
          type: ["string", "null"]
        pageCount:
          type: integer
        totalPriceLek:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        pages:
          type: array
          items:
            $ref: "#/components/schemas/ProjectPage"
    AutoSaveInput:
      type: object
      required: [pagesJson]
      properties:
        pagesJson:
          type: string
          description: JSON string of all page states
    PdfGenerationResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, generating, ready, error]
        pdfUrl:
          type: ["string", "null"]
        shareToken:
          type: ["string", "null"]
        error:
          type: ["string", "null"]

    # ── Orders ────────────────────────────────────────────────────────────
...